Revert "Levenshtein distance 2 is implemented and an English word dataset is …"
