+++ .gitignore
@@ -1,3 +1,2 @@
 # Ignore the virtual environment folder
 spelling-correction_env/
-test.py
